+++ src/common/keycodes.ts
@@ -1,18 +1,16 @@
-export const UP_ARROW = 38;
-export const DOWN_ARROW = 40;
-export const RIGHT_ARROW = 39;
-export const LEFT_ARROW = 37;
-
-export const PAGE_UP = 33;
-export const PAGE_DOWN = 34;
-
-export const HOME = 36;
-export const END = 35;
-
-export const ENTER = 13;
-export const SPACE = 32;
-export const TAB = 9;
-
-export const ESCAPE = 27;
-export const BACKSPACE = 8;
-export const DELETE = 46;
+export enum KeyCode {
+    UpArrow = 38,
+    DownArrow = 40,
+    RightArrow = 39,
+    LeftArrow = 37,
+    PageUp = 33,
+    PageDown = 34,
+    Home = 36,
+    End = 35,
+    Enter = 13,
+    Space = 32,
+    Tab = 9,
+    Escape = 27,
+    Backspace = 8,
+    Delete = 46
+}